+++ Tests/Stopping.test.ps1
@@ -1,14 +1,10 @@
-
 <#
 .Synopsis
 	Tests stopping of Split-Pipeline.
-
-.Link
-	Invoked by https://github.com/nightroman/Invoke-Build
 #>
 
-Import-Module SplitPipeline
-Set-StrictMode -Version Latest
+#requires -Modules SplitPipeline
+Set-StrictMode -Version 3
 
 <#
 	[Ctrl-C] hangs in v1.2.0, works in 1.2.1
@@ -25,12 +21,12 @@ Set-StrictMode -Version Latest
 	waits for them.
 
 	- Press [Ctrl-C] in the calling console. Split-Pipeline still waits because
-	WaitForExit is not stopped so, I guess.
+	WaitForExit is not stopped this way.
 
 	- Close notepads. Split-Pipeline exits, not hangs.
 #>
 task Issue3 {
-	assert (!(Get-Process [n]otepad))
+	assert (!(Get-Process wordpad -ErrorAction Ignore))
 
 	remove C:\TEMP\SplitPipelineIssue3
 	$null = mkdir C:\TEMP\SplitPipelineIssue3
@@ -41,7 +37,7 @@ task Issue3 {
 		Import-Module SplitPipeline
 		1..4 | Split-Pipeline -Verbose -Count 2 `
 		-Script {process{
-			$p = Start-Process notepad -PassThru
+			$p = Start-Process wordpad -PassThru
 			$p.WaitForExit()
 		}} `
 		-Begin {
@@ -60,23 +56,23 @@ task Issue3 {
 	'BeginInvoke'
 	$null = $ps.BeginInvoke()
 
-	# wait for two jobs to start, i.e. two notepads
-	while(@(Get-Process [n]otepad).Count -lt 2) {
+	# wait for two jobs to start, i.e. two processes
+	while(@(Get-Process wordpad -ErrorAction Ignore).Count -lt 2) {
 		Start-Sleep -Milliseconds 100
 	}
 
 	# 2 jobs started
-	equals @(Get-Process [n]otepad).Count 2
+	equals @(Get-Process wordpad).Count 2
 
 	# start stopping, fake [Ctrl-C]
 	'BeginStop'
 	$a2 = $ps.BeginStop($null, $null)
 
-	#! kill notepads, this releases jobs
+	#! kill processes, this releases jobs
 	#! PSv2 Stop-Process is not enough
 	Start-Sleep 2
-	while(Get-Process [n]otepad) {
-		Stop-Process -Name [n]otepad
+	while(Get-Process wordpad -ErrorAction Ignore) {
+		Stop-Process -Name wordpad
 		Start-Sleep -Milliseconds 100
 	}
 
@@ -84,9 +80,9 @@ task Issue3 {
 	'WaitOne'
 	$null = $a2.AsyncWaitHandle.WaitOne()
 
-	# no new jobs or notepads (3 and 4)
+	# no new jobs or processes (3 and 4)
 	Start-Sleep 2
-	assert (!(Get-Process [N]otepad))
+	assert (!(Get-Process wordpad -ErrorAction Ignore))
 
 	# logs
 	$logs = Get-Item C:\TEMP\SplitPipelineIssue3\*