+++ Tests/Test-Refill.ps1
@@ -34,7 +34,7 @@
 	Besides, order of Split-Pipeline results is not necessarily constant.
 
 .Link
-	https://github.com/nightroman/SplitPipeline/blob/master/Tests/Test-Refill.ps1
+	https://github.com/nightroman/SplitPipeline/blob/main/Tests/Test-Refill.ps1
 #>
 
 ### Hierarchical data: container nodes are represented by hashtables