+++ Tests/Test-Start-Job.ps1
@@ -6,8 +6,8 @@
 	Use Start-Job to run pipelines in separate processes, e.g. in cases like
 	https://github.com/nightroman/SplitPipeline/issues/32
 
-	The sample jobs would take ~20 seconds when run sequentially.
-	With Split-Pipeline and Start-Job they take ~7 seconds.
+	The sample jobs would take ~8 seconds when run sequentially.
+	With Split-Pipeline and Start-Job they take ~4 seconds.
 
 	Note that Start-Job is relatively expensive and
 	Split-Pipeline may work slower with faster jobs.
@@ -17,7 +17,7 @@ Import-Module SplitPipeline
 
 $sw = [System.Diagnostics.Stopwatch]::StartNew()
 
-$data = 1..20 | Split-Pipeline -Count 4 {process{
+$data = 1..8 | Split-Pipeline -Count 4 {process{
 	$job = Start-Job -ArgumentList $_ {
 		# fake time consuming job
 		Start-Sleep 1